backend/requirements.txt
fastapi==0.94.1
uvicorn==0.21.1
asyncpg==0.27.0
tortoise-orm==0.19.3
aerich==0.7.1
httpx==0.23.3
pytest==7.2.2
pandas==2.0.1
